+++ backend/src/main/java/com/pickpick/channel/application/ChannelService.java
@@ -1,6 +1,6 @@
-package com.pickpick.service;
+package com.pickpick.channel.application;
 
-import com.pickpick.repository.ChannelRepository;
+import com.pickpick.channel.domain.ChannelRepository;
 import org.springframework.stereotype.Service;
 
 @Service
