+++ backend/src/main/java/com/pickpick/channel/domain/Channel.java
@@ -1,8 +1,12 @@
-package com.pickpick.entity;
+package com.pickpick.channel.domain;
 
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
 import lombok.Getter;
-
-import javax.persistence.*;
 
 @Getter
 @Table(name = "channel")
