+++ backend/src/main/java/com/pickpick/channel/ui/dto/ChannelSubscriptionRequest.java
@@ -1,4 +1,4 @@
-package com.pickpick.controller.dto;
+package com.pickpick.channel.ui.dto;
 
 import lombok.Getter;
