+++ backend/src/test/java/com/pickpick/acceptance/ChannelAcceptanceTest.java
@@ -1,17 +1,16 @@
 package com.pickpick.acceptance;
 
-import com.pickpick.controller.dto.ChannelResponse;
-import com.pickpick.controller.dto.ChannelSubscriptionRequest;
+import static org.assertj.core.api.Assertions.assertThat;
+
+import com.pickpick.channel.ui.dto.ChannelResponse;
+import com.pickpick.channel.ui.dto.ChannelSubscriptionRequest;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
+import java.util.List;
+import java.util.stream.Collectors;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.test.context.jdbc.Sql;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
-import static org.assertj.core.api.Assertions.assertThat;
 
 @Sql({"/truncate.sql", "/channel.sql"})
 @DisplayName("채널 기능")
