+++ backend/src/test/java/com/pickpick/service/MessageDeletedServiceTest.java
@@ -2,10 +2,10 @@ package com.pickpick.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.pickpick.entity.Channel;
+import com.pickpick.channel.domain.Channel;
+import com.pickpick.channel.domain.ChannelRepository;
 import com.pickpick.entity.Member;
 import com.pickpick.entity.Message;
-import com.pickpick.repository.ChannelRepository;
 import com.pickpick.repository.MemberRepository;
 import com.pickpick.repository.MessageRepository;
 import java.time.LocalDateTime;
